# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for user management",
        "title": "User API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/social_media": {
            "get": {
                "description": "Retrieve all social media details from MongoDB",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "SocialMediaDetails"
                ],
                "summary": "Get all social media details",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/SocialMedia.SocialMedia"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new social media detail in MongoDB",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "SocialMediaDetails"
                ],
                "summary": "Create a new social media detail",
                "parameters": [
                    {
                        "description": "Social media detail object",
                        "name": "socialMediaDetail",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/SocialMedia.SocialMedia"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "ID of the created social media detail",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete all social media details from MongoDB",
                "tags": [
                    "SocialMediaDetails"
                ],
                "summary": "Delete all social media details",
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        },
        "/social_media/search": {
            "get": {
                "description": "Search social media details in MongoDB based on a keyword",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "SocialMediaDetails"
                ],
                "summary": "Search social media details",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Search keyword",
                        "name": "keyword",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/SocialMedia.SocialMedia"
                            }
                        }
                    }
                }
            }
        },
        "/social_media/{id}": {
            "get": {
                "description": "Retrieve a social media detail by its ID from MongoDB",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "SocialMediaDetails"
                ],
                "summary": "Get a social media detail by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Social media detail ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/SocialMedia.SocialMedia"
                        }
                    }
                }
            },
            "put": {
                "description": "Update an existing social media detail in MongoDB",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "SocialMediaDetails"
                ],
                "summary": "Update a social media detail",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Social media detail ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Updated social media detail object",
                        "name": "socialMediaDetail",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/SocialMedia.SocialMedia"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            },
            "delete": {
                "description": "Delete a social media detail by its ID from MongoDB",
                "tags": [
                    "SocialMediaDetails"
                ],
                "summary": "Delete a social media detail",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Social media detail ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        }
    },
    "definitions": {
...
        "SocialMedia.SocialMedia": {
            "type": "object",
            "required": [
                "sm_cv",
                "sm_email",
                "sm_facebook",
                "sm_github",
                "sm_instagram",
                "sm_linkedin",
                "sm_twitter",
                "sm_whatsapp"
            ],
            "properties": {
                "sm_cv": {
                    "type": "string",
                    "example": "ahmed_mady_cv.pdf"
                },
                "sm_email": {
                    "type": "string",
                    "example": "ahmed.mady@example.com"
                },
                "sm_facebook": {
                    "type": "string",
                    "example": "ahmed.mady.facebook"
                },
                "sm_github": {
                    "type": "string",
                    "example": "ahmed_mady_github"
                },
                "sm_id": {
                    "type": "string"
                },
                "sm_instagram": {
                    "type": "string",
                    "example": "ahmed_mady_instagram"
                },
                "sm_linkedin": {
                    "type": "string",
                    "example": "john-doe-linkedin"
                },
                "sm_twitter": {
                    "type": "string",
                    "example": "@ahmed_mady_twitter"
                },
                "sm_whatsapp": {
                    "type": "string",
                    "example": "123456789"
                }
            }
        }
    }
}